+++ packages/components/scripts/vendorism/custom/index.js
@@ -1,3 +1,4 @@
+export * from './alert.vendorism.js';
 export * from './badge.vendorism.js';
 export * from './button.vendorism.js';
 export * from './drawer.vendorism.js';
@@ -6,6 +7,7 @@ export * from './icon.vendorism.js';
 export * from './input.vendorism.js';
 export * from './menu-item.vendorism.js';
 export * from './menu-label.vendorism.js';
+export * from './progress-bar.vendorism.js';
 export * from './select.vendorism.js';
 export * from './tag.vendorism.js';
 export * from './textarea.vendorism.js';