+++ packages/components/src/components/side-nav/side-nav.component.ts
@@ -51,7 +51,6 @@ import { unlockBodyScrolling } from '../../internal/scroll.js';
  * @csspart overlay - The overlay that covers the screen behind the side-nav.
  *
  * @cssproperty  --side-nav-open-width - The width of the side-nav if in open state
- * @cssproperty  --side-nav-rail-width - The width of the side-nav if in rail mode without open
  *
  * @animation sideNav.showNonRail - The animation to use when showing the side-nav in non-rail mode.
  * @animation sideNav.showRail - The animation to use when showing the side-nav in rail mode.
@@ -332,8 +331,8 @@ export default class SynSideNav extends SynergyElement {
 // Show animations
 setDefaultAnimation('sideNav.showRail', {
   keyframes: [
-    { width: '4.5rem' },
-    { width: '25rem' },
+    { width: 'var(--side-nav-rail-width)' },
+    { width: 'var(--side-nav-open-width)' },
   ],
   options: { duration: 250, easing: 'ease' },
 });
@@ -357,8 +356,8 @@ setDefaultAnimation('sideNav.hideNonRail', {
 
 setDefaultAnimation('sideNav.hideRail', {
   keyframes: [
-    { width: '25rem' },
-    { width: '4.5rem' },
+    { width: 'var(--side-nav-open-width)' },
+    { width: 'var(--side-nav-rail-width)' },
   ],
   options: { duration: 250, easing: 'ease' },
 });