+++ packages/components/web-test-runner.config.js
@@ -31,7 +31,11 @@ export default {
   ],
   browsers: [
     playwrightLauncher({ product: 'chromium' }),
-    playwrightLauncher({ product: 'firefox' }),
+    
+    // Enable firefox, but use concurrency of 1
+    // @see https://github.com/modernweb-dev/web/issues/2374
+    playwrightLauncher({ product: 'firefox', concurrency: 1 }),
+
     playwrightLauncher({ product: 'webkit' })
   ],
   testRunnerHtml: testFramework => `