+++ packages/docs/stories/components/alert.stories.ts
@@ -1,0 +1,339 @@
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable import/no-relative-packages */
+import type { Meta, StoryObj } from '@storybook/web-components';
+import { html } from 'lit';
+import '../../../components/src/components/alert/alert.js';
+import {
+  generateScreenshotStory,
+  generateStoryDescription,
+  storybookDefaults,
+  storybookHelpers,
+  storybookTemplate,
+} from '../../src/helpers/component.js';
+import { generateFigmaPluginObject } from '../../src/helpers/figma.js';
+
+const { args: defaultArgs, argTypes } = storybookDefaults('syn-alert');
+const { overrideArgs } = storybookHelpers('syn-alert');
+const { generateTemplate } = storybookTemplate('syn-alert');
+
+const meta: Meta = {
+  args: overrideArgs([
+    {
+      name: 'default',
+      type: 'slot',
+      value: 'This is a standard alert. You can customize its content and even the icon.',
+    },
+    {
+      name: 'open',
+      type: 'attribute',
+      value: true,
+    },
+    {
+      name: 'icon',
+      type: 'slot',
+      value: '<syn-icon slot="icon" name="info"></syn-icon>',
+    },
+  ], defaultArgs),
+  argTypes,
+  component: 'syn-alert',
+  parameters: {
+    design: generateFigmaPluginObject('15172-32035'),
+    docs: {
+      description: {
+        component: generateStoryDescription('alert', 'default'),
+      },
+    },
+  },
+  title: 'Components/syn-alert',
+};
+export default meta;
+
+type Story = StoryObj;
+
+export const Default: Story = {
+  parameters: {
+    controls: {
+      disable: false,
+    },
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'default'),
+      },
+    },
+  },
+  render: (args: unknown) => generateTemplate({ args }),
+} as Story;
+
+export const Variants: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'variants'),
+      },
+    },
+  },
+  render: () => html`
+    <div style="display: flex; flex-direction: column; gap: var(--syn-spacing-medium);">
+      <syn-alert variant="primary" open>
+        <syn-icon slot="icon" name="info"></syn-icon>
+        <strong>This is super informative</strong><br />
+        You can tell by how pretty the alert is.
+      </syn-alert>
+
+      <syn-alert variant="success" open>
+        <syn-icon slot="icon" name="check_circle"></syn-icon>
+        <strong>Your changes have been saved</strong><br />
+        You can safely exit the app now.
+      </syn-alert>
+
+      <syn-alert variant="neutral" open>
+        <syn-icon slot="icon" name="settings"></syn-icon>
+        <strong>Your settings have been updated</strong><br />
+        Settings will take effect on next login.
+      </syn-alert>
+
+      <syn-alert variant="warning" open>
+        <syn-icon slot="icon" name="warning"></syn-icon>
+        <strong>Your session has ended</strong><br />
+        Please login again to continue.
+      </syn-alert>
+
+      <syn-alert variant="danger" open>
+        <syn-icon slot="icon" name="error"></syn-icon>
+        <strong>Your account has been deleted</strong><br />
+        We're very sorry to see you go!
+      </syn-alert>
+    </div>
+  `,
+};
+
+export const Closable: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'closable'),
+      },
+    },
+  },
+  render: () => html`
+    <syn-alert variant="primary" open closable class="alert-closable">
+      <syn-icon slot="icon" name="info"></syn-icon>
+      You can close this alert any time!
+    </syn-alert>
+    
+    <script type="module">
+    const alert = document.querySelector('.alert-closable');
+    alert.addEventListener('syn-after-hide', () => {
+      setTimeout(() => (alert.open = true), 2000);
+    });
+    </script>
+  `,
+};
+
+export const WithoutIcons: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'without-icons'),
+      },
+    },
+  },
+  render: () => html`
+    <syn-alert variant="primary" open>
+      Nothing fancy here, just a simple alert.
+    </syn-alert>
+  `,
+};
+
+export const Duration: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'duration'),
+      },
+    },
+  },
+  render: () => html`
+    <div class="alert-duration">
+      <syn-button variant="outline">Show Alert</syn-button>
+
+      <syn-alert variant="primary" duration="3000" closable>
+        <syn-icon slot="icon" name="info"></syn-icon>
+        This alert will automatically hide itself after three seconds, unless you interact with it.
+      </syn-alert>
+    </div>
+
+    <script type="module">
+    const container = document.querySelector('.alert-duration');
+    const button = container.querySelector('syn-button');
+    const alert = container.querySelector('syn-alert');
+
+    button.addEventListener('click', () => alert.show());
+    </script>
+
+    <style>
+    .alert-duration syn-alert {
+      margin-top: var(--syn-spacing-small);
+    }
+    </style>
+  `,
+};
+
+export const ToastNotifications: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'toast-notifications'),
+      },
+    },
+  },
+  render: () => html`
+    <div class="alert-toast">
+      <div style="display: flex; gap: var(--syn-spacing-small);">
+        <syn-button data-variant="primary">Primary</syn-button>
+        <syn-button data-variant="success">Success</syn-button>
+        <syn-button data-variant="neutral">Neutral</syn-button>
+        <syn-button data-variant="warning">Warning</syn-button>
+        <syn-button data-variant="danger">Danger</syn-button>
+      </div>
+
+      <syn-alert variant="primary" duration="3000" closable>
+        <syn-icon slot="icon" name="info"></syn-icon>
+        <strong>This is super informative</strong><br />
+        You can tell by how pretty the alert is.
+      </syn-alert>
+
+      <syn-alert variant="success" duration="3000" closable>
+        <syn-icon slot="icon" name="check_circle"></syn-icon>
+        <strong>Your changes have been saved</strong><br />
+        You can safely exit the app now.
+      </syn-alert>
+
+      <syn-alert variant="neutral" duration="3000" closable>
+        <syn-icon slot="icon" name="settings"></syn-icon>
+        <strong>Your settings have been updated</strong><br />
+        Settings will take effect on next login.
+      </syn-alert>
+
+      <syn-alert variant="warning" duration="3000" closable>
+        <syn-icon slot="icon" name="warning"></syn-icon>
+        <strong>Your session has ended</strong><br />
+        Please login again to continue.
+      </syn-alert>
+
+      <syn-alert variant="danger" duration="3000" closable>
+        <syn-icon slot="icon" name="error"></syn-icon>
+        <strong>Your account has been deleted</strong><br />
+        We're very sorry to see you go!
+      </syn-alert>
+    </div>
+
+    <script type="module">
+    const container = document.querySelector('.alert-toast');
+
+    ['primary', 'success', 'neutral', 'warning', 'danger'].map(variant => {
+      const button = container.querySelector(\`syn-button[data-variant="\${variant}"]\`);
+      const alert = container.querySelector(\`syn-alert[variant="\${variant}"]\`);
+
+      button.addEventListener('click', () => alert.toast());
+    });
+    </script>
+  `,
+};
+
+export const CreatingToastsImperatively: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('alert', 'creating-toasts-imperatively'),
+      },
+    },
+  },
+  render: () => html`
+    <div class="alert-toast-wrapper">
+      <syn-button>Create Toast</syn-button>
+    </div>
+
+    <script type="module">
+    const container = document.querySelector('.alert-toast-wrapper');
+    const button = container.querySelector('syn-button');
+    let count = 0;
+
+    // Always escape HTML for text arguments!
+    function escapeHtml(html) {
+      const div = document.createElement('div');
+      div.textContent = html;
+      return div.innerHTML;
+    }
+
+    // Custom function to emit toast notifications
+    function notify(message, variant = 'primary', icon = 'info', duration = 3000) {
+      const alert = Object.assign(document.createElement('syn-alert'), {
+        variant,
+        closable: true,
+        duration: duration,
+        innerHTML: \`
+          <syn-icon name="\${icon}" slot="icon"></syn-icon>
+          This alert will automatically hide itself after three seconds, unless you interact with it.
+        \`,
+      });
+
+      document.body.append(alert);
+      return alert.toast();
+    }
+
+    button.addEventListener('click', () => {
+      notify(\`This is custom toast #\${++count}\`);
+    });
+    </script>
+  `,
+};
+
+// This story is only there to make screenshots comparable
+// and therefore has no description or controls
+const Toast = {
+  render: () => html`
+    <style>
+    .syn-toast-stack {
+      position: relative;
+    }
+    </style>
+    <div class="syn-toast-stack">
+      <syn-alert variant="primary" open closable>
+        <syn-icon slot="icon" name="info"></syn-icon>
+        This is super informative
+      </syn-alert>
+
+      <syn-alert variant="success" open closable>
+        <syn-icon slot="icon" name="check_circle"></syn-icon>
+        Your changes have been saved
+      </syn-alert>
+
+      <syn-alert variant="neutral" open closable>
+        <syn-icon slot="icon" name="settings"></syn-icon>
+        Your settings have been updated
+      </syn-alert>
+
+      <syn-alert variant="warning" open closable>
+        <syn-icon slot="icon" name="warning"></syn-icon>
+        Your session has ended
+      </syn-alert>
+
+      <syn-alert variant="danger" open closable>
+        <syn-icon slot="icon" name="error"></syn-icon>
+        Your account has been deleted
+      </syn-alert>
+    </div>
+  `,
+};
+
+/* eslint-disable sort-keys */
+export const Screenshot: Story = generateScreenshotStory({
+  Default,
+  Variants,
+  Closable,
+  WithoutIcons,
+  Toast,
+}, 550);
+/* eslint-enable sort-keys */